Project restructure.
