+++ app/src/main/java/com/github/featuredetectandroid/ui/OutputViewModel.kt
@@ -6,15 +6,15 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.geometry.Offset
 import androidx.lifecycle.ViewModel
-import com.github.featuredetectandroid.utils.keypointsToOffsetList
-import com.github.featuredetectandroid.utils.luminanceArrayToBitmap
+import com.github.featuredetectandroid.utils.conversions.keypointsToOffsetList
+import com.github.featuredetectandroid.utils.conversions.luminanceArrayToBitmap
 import com.github.featuredetectlib.FeatureDetector
 import com.github.featuredetectlib.Keypoint
 
-class GrayscaleViewModel : ViewModel() {
+class OutputViewModel : ViewModel() {
     var grayscaleBitmap: Bitmap? by mutableStateOf(null)
     var isCameraPermissionGranted by mutableStateOf(false)
-    var keypointsOffset: List<Offset>? by mutableStateOf(null)
+    var keypointsOffset: List<Offset> by mutableStateOf(emptyList())
     var featureDetector: FeatureDetector? by mutableStateOf(null)
     private var width by mutableStateOf(0)
     private var height by mutableStateOf(0)
@@ -25,7 +25,7 @@ class GrayscaleViewModel : ViewModel() {
         grayscaleBitmap = luminanceArrayToBitmap(grayscaleByteArray, newWidth, newHeight)
     }
 
-    fun setKeypoints(keypoints: List<Keypoint>) {
+    fun setKeypointsForOutput(keypoints: List<Keypoint>) {
         keypointsOffset = keypointsToOffsetList(keypoints)
     }
