+++ app/src/main/java/com/github/featuredetectandroid/utils/conversions/KeypointsToOffsetList.kt
@@ -1,4 +1,4 @@
-package com.github.featuredetectandroid.utils
+package com.github.featuredetectandroid.utils.conversions
 
 import androidx.compose.ui.geometry.Offset
 import com.github.featuredetectlib.Keypoint
